document-agent/requirements.txt
aiohttp==3.11.16
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
boto3==1.37.27
botocore==1.37.27
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
diskcache==5.6.3
docstring_parser==0.16
emoji==2.14.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.87.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gql==3.5.2
graphql-core==3.2.4
grpcio==1.71.0
grpcio-status==1.71.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
instructor==1.7.7
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
langchain==0.3.22
langchain-aws==0.2.18
langchain-community==0.3.20
langchain-core==0.3.50
langchain-ollama==0.3.0
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.24
litellm
markdown-it-py==3.0.0
marshmallow==3.26.1
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.7
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
SQLAlchemy==2.0.40
streamlit==1.44.1
tenacity==9.1.2
tiktoken==0.9.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.13.1
urllib3==2.3.0
wandb==0.19.9
weave==0.51.41
xxhash==3.5.0
nest-asyncio==1.6.0

# Document processing dependencies
# RT-DETR and computer vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# OCR and text processing
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=3.1.0
fitz>=0.0.1.dev2

# Image processing and layout detection
scikit-image>=0.21.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Document analysis and NLP
spacy>=3.6.0
nltk>=3.8.0
transformers>=4.30.0
accelerate>=0.20.0

# Additional utilities
tqdm>=4.65.0
colorama>=0.4.6
tabulate>=0.9.0